Refactor to a loop-based usb-client
Basically we have to read data that comes in on the serial bus to make
the USB host happy (macOS actually freezes for a second periodically
otherwise). If we do that based on a registered interrupt, we have to
use the `interrupt` macro, which makes it a pain to inject
dependencies for our handler function, even as global mutable
variables. The alternative is to do it as part of the loop, which
works fine as long as the loop doesn't block too long for other
reasons.